# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth & Authen API",
        "description": "API documentation for Authentication endpoints",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/email/verification-notification": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend email verification link",
                "operationId": "d5d22620f2dd4f8bd7761f60f5347166",
                "responses": {
                    "200": {
                        "description": "Verification email sent successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify a user's email address",
                "description": "Handles email verification requests. If the email is already verified, it returns a relevant message or redirects to the frontend. If not verified, it fulfills verification and redirects or responds with JSON.",
                "operationId": "verifyEmail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user whose email is being verified",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "The email verification hash",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abcdef1234567890"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified or already verified (JSON response)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Email verified successfully."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Email already verified."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "302": {
                        "description": "Redirects to frontend confirmation page with status parameter",
                        "headers": {
                            "Location": {
                                "description": "Frontend redirect URL (e.g., /email-verified?status=success or /email-verified?status=already_verified)",
                                "schema": {
                                    "type": "string",
                                    "example": "https://your-frontend.com/email-verified?status=success"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or expired verification link"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/email/generate-verification-url": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Generate a temporary email verification link for the authenticated user",
                "operationId": "c4623e90839a40ae936337bae37feb4d",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "verification_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/api/verify-email/1/9b7c7f03b3f75f7d1bb5cbfd7ad9f5b7f056882c?expires=1735300078&signature=..."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Email already verified."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}